+++ packages/plugin-hardhat/src/index.ts
@@ -30,4 +30,4 @@ export type { HardhatUpgrades, DefenderHardhatUpgrades } from './types.js';
 // Utilities
 export type { UpgradeOptions } from './utils/options.js';
 
-export { OZfilesToBuild } from './utils/npmFilesToBuild.js';
+export { proxyFilesToBuild } from './utils/npmFilesToBuild.js';